message_pact_python/requirements.txt
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-mock==3.7.0
pact-python==1.6.0